+++ CHANGELOG.md
@@ -5,15 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.0-alpha-1] - 2023-10-10
 
 ## [0.2.0] - 2023-10-08
 
 ### Added
 
+- Added prototype library page.
+- Added serving of static files.
+
 ### Fixed
 
 ### Changed
+
+- Refactored pages to use askama templates.
 
 ### Removed